backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.46
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic==2.4.2
pydantic-settings==2.1.0
aiosqlite==0.19.0
httpx==0.25.2
openai>=1.10.0
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10